requirements.txt
ipython==0.13.1
ipdb==0.7
Sphinx==1.1.3
mock==1.0.1
cassandra-driver